% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_gene_sets}
\alias{build_gene_sets}
\title{Build gene sets from marker tables}
\usage{
build_gene_sets(files, dir = ".", gene_col = "gene")
}
\arguments{
\item{files}{Named character vector of marker files.}

\item{dir}{Directory containing files.}

\item{gene_col}{Gene column.}
}
\value{
Named list of gene sets.
}
\description{
Build gene sets from marker tables
}
