% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cellpal}
\alias{cellpal}
\title{Generate plotting palette}
\usage{
cellpal(n, preset = c("base", "bright", "pastel"), reverse = FALSE)
}
\arguments{
\item{n}{Number of colors.}

\item{preset}{Palette preset.}

\item{reverse}{Logical; whether to reverse color order.}
}
\value{
A character vector of colors.
}
\description{
Generate plotting palette
}
