% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_expression_matrix}
\alias{extract_expression_matrix}
\title{Extract expression matrix}
\usage{
extract_expression_matrix(object, assay = "RNA", layer = "counts")
}
\arguments{
\item{object}{Seurat object.}

\item{assay}{Assay name.}

\item{layer}{Data layer.}
}
\value{
Sparse expression matrix.
}
\description{
Extract expression matrix
}
