% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-diagnostics.R
\name{gnrh_cellcount}
\alias{gnrh_cellcount}
\title{Plot GnRH classification counts}
\usage{
gnrh_cellcount(object, txtsize = getOption("gnrhcell.base_size", 14))
}
\arguments{
\item{object}{A Seurat object processed by \code{run_gnrh()}.}

\item{txtsize}{Base text size.}
}
\value{
A ggplot object.
}
\description{
Plot GnRH classification counts
}
