% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-theme.R
\name{gnrh_palette}
\alias{gnrh_palette}
\title{Color palettes}
\usage{
gnrh_palette(
  type = c("status", "confident", "class", "stage", "developmental", "secretory")
)
}
\arguments{
\item{type}{Palette type.}
}
\value{
Named character vector.
}
\description{
Color palettes
}
