% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hpsc}
\alias{hpsc}
\title{Human hPSC GnRH demo dataset}
\format{
A Seurat object.
}
\source{
Wang et al. 2022.
}
\description{
A sampled Seurat object derived from the public Wang et al. 2022
hPSC-derived GnRH neuron dataset.
}
\details{
The object is intended for examples, tutorials, and testing of the
GnRHcell workflow.
}
