% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpsc_gnrh}
\alias{hpsc_gnrh}
\title{Example GnRHcell Seurat object}
\format{
A Seurat object with 500 cells.
}
\source{
Derived from a small subset of a public scRNA-seq dataset.
}
\usage{
hpsc_gnrh
}
\description{
A small example Seurat object for demonstrating GnRHcell functions.
}
\examples{
data(hpsc_gnrh)
hpsc_gnrh

}
\keyword{datasets}
