% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_gene_symbols}
\alias{match_gene_symbols}
\title{Match gene symbols ignoring case}
\usage{
match_gene_symbols(target, pool)
}
\arguments{
\item{target}{Character vector of target genes.}

\item{pool}{Character vector of available genes.}
}
\value{
Matched gene symbols.
}
\description{
Match gene symbols ignoring case
}
