% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_class_counts}
\alias{plot_class_counts}
\title{Plot GnRH classification counts}
\usage{
plot_class_counts(object, txtsize = 9)
}
\arguments{
\item{object}{A Seurat object processed by \code{run_gnrh()}.}

\item{txtsize}{Base text size.}
}
\value{
A ggplot object.
}
\description{
Plot GnRH classification counts
}
