% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gnrh_embedding}
\alias{plot_gnrh_embedding}
\title{DimPlot Wrapper}
\usage{
plot_gnrh_embedding(
  obj,
  group.by = "all",
  reduction = "umap",
  cols = NULL,
  label = FALSE,
  label.face = "plain",
  dark = FALSE
)
}
\arguments{
\item{obj}{Seurat object.}

\item{group.by}{Metadata column used for grouping.}

\item{reduction}{Dimensional reduction to use.}

\item{cols}{Optional colors.}

\item{label}{Logical; draw labels.}

\item{label.face}{Font face for labels.}

\item{dark}{Logical; use dark theme.}
}
\description{
DimPlot Wrapper
}
