% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gnrh_hits}
\alias{plot_gnrh_hits}
\title{Plot GnRH module hit distributions}
\usage{
plot_gnrh_hits(
  data,
  x,
  fill,
  palette = NULL,
  type = c("count", "fraction"),
  title = NULL,
  txtsize = 10
)
}
\description{
Plot GnRH module hit distributions
}
