% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gnrh_roc}
\alias{plot_gnrh_roc}
\title{Plot ROC curve}
\usage{
plot_gnrh_roc(object)
}
\arguments{
\item{object}{A Seurat object containing \code{gnrh_truth_proxy}
and \code{gnrh_expr} metadata fields.}
}
\value{
A ggplot object.
}
\description{
Computes and visualizes ROC performance for GnRH classification.
}
