% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gnrh_specificity}
\alias{plot_gnrh_specificity}
\title{Plot GnRH specificity landscape}
\usage{
plot_gnrh_specificity(object, txtsize = 9)
}
\arguments{
\item{object}{A Seurat object processed by \code{run_gnrh()}.}

\item{txtsize}{Base text size.}
}
\value{
A ggplot object.
}
\description{
Plot GnRH specificity landscape
}
