% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-utils.R
\name{prepare_gnrh_datasets}
\alias{prepare_gnrh_datasets}
\title{Prepare a GnRHcell dataset configuration table}
\usage{
prepare_gnrh_datasets(datasets, check_files = TRUE, remove_missing = FALSE)
}
\arguments{
\item{datasets}{Data frame describing datasets to be processed by the
multi-dataset GnRHcell workflow.}

\item{check_files}{Logical. Check whether dataset files exist.}

\item{remove_missing}{Logical. Remove datasets whose input files cannot be
found.}
}
\description{
Prepare a GnRHcell dataset configuration table
}
