% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{print.gnrh_markers}
\alias{print.gnrh_markers}
\title{Print GnRH marker results}
\usage{
\method{print}{gnrh_markers}(x, ...)
}
\arguments{
\item{x}{A gnrh_markers object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns input object
}
\description{
Print GnRH marker results
}
