% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_staging.R
\name{stage_gnrh_cells}
\alias{stage_gnrh_cells}
\title{Stage GnRH-expressing cells}
\usage{
stage_gnrh_cells(object, assay = "RNA", layer = "counts", verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object.}

\item{assay}{Assay name.}

\item{layer}{Expression layer.}

\item{verbose}{Print progress messages.}
}
\value{
Updated Seurat object.
}
\description{
Assigns developmental stages using predefined
GnRH lineage gene programs.
}
