% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_gnrh_run_info}
\alias{extract_gnrh_run_info}
\title{Extract GnRHcell run information}
\usage{
extract_gnrh_run_info(object, dataset_name = NULL)
}
\arguments{
\item{object}{A Seurat object processed with \code{run_gnrh()}.}

\item{dataset_name}{Optional dataset name.}
}
\value{
One-row data frame of GnRHcell run statistics.
}
\description{
Extract GnRHcell run information
}
