% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{gnrh_colors}
\alias{gnrh_colors}
\title{GnRHcell color palettes}
\usage{
gnrh_colors(type = c("status", "confident", "class", "stage", "secretory"))
}
\arguments{
\item{type}{Palette type.}
}
\value{
Named character vector.
}
\description{
GnRHcell color palettes
}
