% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gnrh_gene_upset}
\alias{gnrh_gene_upset}
\title{Gene set overlap analysis and visualization}
\usage{
gnrh_gene_upset(
  gene_sets,
  min_size = 1,
  venn_title = "Overlap of Gene Sets",
  outdir = ".",
  save_plot = TRUE,
  max_intersections = Inf,
  plot_width = NULL,
  plot_height = NULL,
  dpi = 600
)
}
\description{
Gene set overlap analysis and visualization
}
