% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker-programs.R
\name{gnrh_secretory_marker_module}
\alias{gnrh_secretory_marker_module}
\title{GnRH secretory marker program}
\usage{
gnrh_secretory_marker_module()
}
\value{
Character vector of secretory genes.
}
\description{
GnRH secretory marker program
}
