% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker-programs.R
\name{gnrh_stage_modules}
\alias{gnrh_stage_modules}
\title{Default GnRH developmental marker programs}
\usage{
gnrh_stage_modules(include_secretory = FALSE)
}
\arguments{
\item{include_secretory}{Include secretory module for backward compatibility.}
}
\value{
Named list of developmental gene programs.
}
\description{
Default GnRH developmental marker programs
}
