% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{GnRHcell-package}
\alias{GnRHcell}
\alias{GnRHcell-package}
\title{GnRHcell: Detection and characterization of GnRH neurons}
\description{
Tools for detecting rare GnRH neurons in single-cell RNA-seq datasets,
developmental staging, diagnostics, marker discovery, and cross-dataset
comparison.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ymbouamboua.github.io/GnRHcell/}
  \item \url{https://github.com/ymbouamboua/GnRHcell}
  \item Report bugs at \url{https://github.com/ymbouamboua/GnRHcell/issues}
}

}
\author{
\strong{Maintainer}: Yvon Mbouamboua \email{yvon.mbouamboua@gmail.com}

}
\keyword{internal}
