% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{GnRHcell-package}
\alias{GnRHcell}
\alias{GnRHcell-package}
\title{GnRHcell package internal imports}
\description{
Provides tools to detect, score, and stage GnRH-like cells in single-cell RNA-seq data using gene modules, adaptive scoring, ambient RNA correction, and Seurat-compatible workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ymbouamboua.github.io/GnRHcell/}
  \item \url{https://github.com/ymbouamboua/GnRHcell}
  \item Report bugs at \url{https://github.com/ymbouamboua/GnRHcell/issues}
}

}
\author{
\strong{Maintainer}: Yvon Mbouamboua \email{yvon.mbouamboua@gmail.com}

}
\keyword{internal}
