% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gnrh_detected}
\alias{plot_gnrh_detected}
\title{Plot detected GnRH-positive cells across datasets}
\usage{
plot_gnrh_detected(
  files = NULL,
  dir = file.path("results", "tables"),
  pattern = "_gnrh_run_info\\\\.tsv$",
  x.ang = 45,
  txtsize = 10,
  debug = FALSE
)
}
\description{
Plot detected GnRH-positive cells across datasets
}
