% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gnrh_distribution}
\alias{plot_gnrh_distribution}
\title{Plot metadata distributions}
\usage{
plot_gnrh_distribution(
  object,
  group.by,
  split.by = NULL,
  cols = NULL,
  sort = FALSE,
  decreasing = TRUE,
  proportion = FALSE,
  position = "stack",
  label = TRUE,
  label.size = 3,
  border = TRUE,
  border.col = "black",
  border.size = 0.2,
  width = 0.7,
  x.lab = NULL,
  y.lab = NULL,
  plot.ttl = NULL,
  txtsize = 10,
  x.ang = 45,
  style = "test",
  flip = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object.}

\item{group.by}{Metadata column to plot.}

\item{split.by}{Optional metadata column for grouped plots
(e.g. sample, batch, condition).}

\item{cols}{Named color vector.}

\item{sort}{Sort bars.}

\item{decreasing}{Sort decreasing.}

\item{proportion}{Plot proportions instead of counts.}

\item{position}{Bar position:
\code{"stack"} or \code{"dodge"}.}

\item{label}{Add labels.}

\item{label.size}{Label text size.}

\item{border}{Draw borders.}

\item{border.col}{Border color.}

\item{border.size}{Border linewidth.}

\item{width}{Bar width.}

\item{x.lab}{X axis label.}

\item{y.lab}{Y axis label.}

\item{plot.ttl}{Plot title.}

\item{txtsize}{Base text size.}

\item{x.ang}{X axis angle.}

\item{style}{Theme style.}

\item{flip}{Flip coordinates.}

\item{...}{Additional arguments passed to \code{plot_theme()}.}
}
\value{
ggplot object.
}
\description{
Publication-ready barplot utility for Seurat metadata.
}
\details{
Supports:
\itemize{
\item Counts or proportions
\item Distribution by sample/group
\item Stacked or dodged bars
\item Horizontal plots
}
}
