% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gnrh_distribution}
\alias{plot_gnrh_distribution}
\title{Plot GnRHcell metadata distributions}
\usage{
plot_gnrh_distribution(
  object,
  group.by,
  split.by = NULL,
  cols = NULL,
  proportion = FALSE,
  position = "stack",
  label = TRUE,
  label.size = 3,
  plot.ttl = NULL,
  txtsize = 10,
  x.ang = 45,
  flip = FALSE
)
}
\description{
Plot GnRHcell metadata distributions
}
