% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gnrh_embedding}
\alias{plot_gnrh_embedding}
\title{Plot GnRH embedding}
\usage{
plot_gnrh_embedding(
  object,
  group_by = "gnrh_status",
  reduction = NULL,
  dims = c(1, 2),
  shuffle = FALSE,
  raster = NULL,
  raster.dpi = c(2048, 2048),
  alpha = 0.9,
  background_alpha = 0.18,
  n.cells = TRUE,
  percentage = FALSE,
  label = FALSE,
  repel = TRUE,
  label.size = 4,
  label.face = "plain",
  cols = NULL,
  axes = TRUE,
  plot.ttl = NULL,
  legend = TRUE,
  leg.ttl = NULL,
  leg.ttl.size = NULL,
  item.size = 3.5,
  leg.pos = "right",
  leg.dir = "vertical",
  leg.size = NULL,
  leg.ncol = NULL,
  item.border = TRUE,
  txtsize = 12,
  pt.size = NULL,
  dark = FALSE,
  total.cells = FALSE,
  style = "classic",
  ...
)
}
\description{
Plot GnRH embedding
}
