% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gnrh_feature}
\alias{plot_gnrh_feature}
\title{Plot GnRH features on an embedding}
\usage{
plot_gnrh_feature(
  object,
  features,
  cols = NULL,
  palette = "gnrh",
  rev.cols = FALSE,
  na.col = "grey90",
  order = FALSE,
  pt.size = NULL,
  reduction = NULL,
  dims = c(1, 2),
  min.cutoff = NULL,
  max.cutoff = NULL,
  raster = NULL,
  raster.dpi = c(2048, 2048),
  split.by = NULL,
  ncol = NULL,
  layer = "data",
  label = FALSE,
  axes = TRUE,
  combine = TRUE,
  blend = FALSE,
  merge.leg = FALSE,
  legend = TRUE,
  plot.ttl = NULL,
  leg.ttl = NULL,
  leg.pos = "right",
  leg.size = 9,
  txtsize = 10,
  dark = FALSE,
  style = "classic",
  ...
)
}
\description{
Plot GnRH features on an embedding
}
