% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gnrh_marker_programs}
\alias{plot_gnrh_marker_programs}
\title{Plot GnRH marker program results}
\usage{
plot_gnrh_marker_programs(
  programs,
  table = c("summary", "high_confidence", "candidate_table"),
  type = c("bar", "dot", "tile"),
  min_genes = 1,
  mode = "light",
  txtsize = 12,
  x.ang = 45
)
}
\description{
Plot GnRH marker program results
}
