% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gnrh_runtime_curve}
\alias{plot_gnrh_runtime_curve}
\title{Plot GnRHcell runtime across datasets}
\usage{
plot_gnrh_runtime_curve(
  files = NULL,
  dir = file.path("results", "tables"),
  pattern = "_gnrh_run_info\\\\.tsv$",
  metric = "total_sec",
  x.ang = 45,
  txtsize = 10,
  show_points = TRUE
)
}
\description{
Plot GnRHcell runtime across datasets
}
