% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{resolve_reduction}
\alias{resolve_reduction}
\title{Resolve an available dimensional reduction}
\usage{
resolve_reduction(
  object,
  reduction = NULL,
  fallback = c("umap", "umap_scvi", "umap.harmony", "umap.rpca", "pca")
)
}
\description{
Resolve an available dimensional reduction
}
\keyword{internal}
